Split `ViewSession` out from `EditSession`, and use it.
